refactor(*): Merge schemas into models.
This merge has the purpose of simplifying discoverability, releasing and revalidating for both repositories [schemas](https://github.com/laboratoria/schemas) and [models](https://github.com/laboratoria/models).

In this new way, we hope we can improve developer experience when we need to add new proposals meant to schemas and models all-in-one.

--
"Things that change for the same reasons and at the same times should be close to each other." CCP principle